+++ packages/experimental/RadioGroup/src/RadioGroup/RadioGroup.styling.ts
@@ -3,8 +3,11 @@ import { Theme, UseStylingOptions, buildProps } from '@fluentui-react-native/fra
 import { defaultRadioGroupTokens } from './RadioGroupTokens';
 import { fontStyles } from '@fluentui-react-native/tokens';
 
+export const radioGroupStates: (keyof RadioGroupTokens)[] = ['disabled'];
+
 export const stylingSettings: UseStylingOptions<RadioGroupProps, RadioGroupSlotProps, RadioGroupTokens> = {
   tokens: [defaultRadioGroupTokens, radioGroupName],
+  states: radioGroupStates,
   slotProps: {
     root: buildProps(() => ({
       style: {
@@ -13,13 +16,33 @@ export const stylingSettings: UseStylingOptions<RadioGroupProps, RadioGroupSlotP
         flexDirection: 'column',
       },
     })),
-    label: buildProps(
+    label: buildProps(() => ({
+      style: {
+        display: 'flex',
+        alignItems: 'center',
+        flexDirection: 'row',
+        alignSelf: 'flex-start',
+      },
+    })),
+    labelText: buildProps(
       (tokens: RadioGroupTokens, theme: Theme) => ({
+        variant: tokens.variant,
         style: {
+          color: tokens.color,
           ...fontStyles.from(tokens, theme),
         },
       }),
-      [...fontStyles.keys],
+      ['color', ...fontStyles.keys],
+    ),
+    required: buildProps(
+      (tokens: RadioGroupTokens) => ({
+        variant: tokens.variant,
+        style: {
+          color: tokens.requiredColor,
+          paddingStart: tokens.requiredPadding,
+        },
+      }),
+      ['requiredColor', 'requiredPadding'],
     ),
   },
 };